fix: the Keep States feature has been added
