+++ workspaces/tables/src/private/types/KeptStateProps.ts
@@ -1,0 +1,5 @@
+type StateTypes = "expandedRows" | "checkedBoxes" | "selectedRows";
+
+type KeptStateProps = Partial<Record<StateTypes, {[key: string]: boolean}[]>>;
+
+export default KeptStateProps;
